+++ src/sampiclyser/entry_points.py
@@ -43,6 +43,12 @@ def cli() -> None:
     pass
 
 
+@click.group()
+def utilities() -> None:
+    """SAMPIClyser command-line utilities interface"""
+    pass
+
+
 cli.add_command(sampiclyser.sampic_convert_script.decode)
 
 
@@ -405,3 +411,61 @@ def plot_channel_hit_rate(
         fig.savefig(output)
     else:
         plt.show()
+
+
+@utilities.command()
+@click.argument('output_file', type=click.Path(exists=False, path_type=Path))
+def generate_example_hitmap_config(output_file):
+    """
+    Generate an example YAML configuration file describing the setup, necessary for hitmap plotting
+    """
+    example_file_data = dict(
+        plot_layout=(1, 2),
+        sensor_types=dict(
+            type1_name=dict(
+                geometry_type="grouped",
+                rows=5,
+                cols=5,
+                ch_to_coords={
+                    1: [(4, 0)],
+                    2: [(4, 1), (4, 2), (3, 1), (3, 2)],
+                    3: [(2, 0), (2, 1)],
+                    4: [(1, 1)],
+                    5: [(0, 0)],
+                    6: [(0, 3), (1, 3), (2, 3)],
+                    7: [(0, 4), (1, 4), (2, 4), (3, 4)],
+                    8: [(2, 2)],
+                    9: [(3, 3)],
+                    10: [(4, 4)],
+                },
+            ),
+        ),
+        sensor_specifications=dict(
+            sensor1_name=dict(
+                datach_to_sensorch={0: 1, 1: 2, 2: 8, 6: 10, 7: 9},
+                sensor_type="type1_name",
+                global_90rotationUnits=2,
+                global_flip=False,
+            ),
+            sensor2_name=dict(
+                datach_to_sensorch={3: 8, 4: 2, 5: 1, 8: 9, 9: 10},
+                sensor_type="type1_name",
+                global_90rotationUnits=1,
+                global_flip=True,
+            ),
+        ),
+    )
+
+    extension = output_file.suffix.lower()
+    if extension in [".json"]:
+        import json
+
+        with open(output_file, 'w') as outfile:
+            json.dump(example_file_data, outfile, indent=2)
+    elif extension in [".yml", ".yaml"]:
+        import yaml
+
+        with open(output_file, 'w') as outfile:
+            yaml.dump(example_file_data, outfile, default_flow_style=False)
+    else:
+        print(f"Unknown config file extension: {extension}")